Changed tickets page filters area
