+++ src/app/pages/tickest/tickest.page.html
@@ -12,24 +12,13 @@
           <ion-icon slot="end" name="chevron-down"></ion-icon>
         </ion-button>
       </ion-buttons>
-      <ion-searchbar class="ion-no-padding ion-padding-vertical" animated></ion-searchbar>
-      <mat-form-field appearance="fill">
-        <mat-label>Month and Year</mat-label>
-        <input matInput [matDatepicker]="dp" [(ngModel)]="date">
-        <mat-hint>MM/YYYY</mat-hint>
-        <mat-datepicker-toggle matSuffix [for]="dp"></mat-datepicker-toggle>
-        <mat-datepicker #dp
-                        startView="multi-year"
-                        (monthSelected)="setMonthAndYear($event, dp)"
-                        panelClass="example-month-picker">
-        </mat-datepicker>
-      </mat-form-field>
+      <!-- <ion-searchbar class="ion-no-padding ion-padding-vertical" animated></ion-searchbar> -->
 
       <ion-buttons class="butons">
-        <ion-button (click)="openFilter()" matTooltipPosition="below" matTooltipClass="tooltip-primary"
+        <!-- <ion-button (click)="openFilter()" matTooltipPosition="below" matTooltipClass="tooltip-primary"
           matTooltip="Filter by" color="primary">
           <ion-icon slot="icon-only" name="options-outline"></ion-icon>
-        </ion-button>
+        </ion-button> -->
         <ion-button color="primary" fill="solid" (click)="openTicketReportDialog()">REPORT</ion-button>
       </ion-buttons>
     </div>
